--- a/runner/file_test.go
+++ b/runner/file_test.go
@@ -2,6 +2,8 @@ package runner
 
 import (
 	"encoding/json"
+	"fmt"
+	"time"
 	"io/ioutil"
 	"os"
 	"testing"
@@ -118,4 +120,22 @@ func Test_transformJSONConcat(t *testing.T) {
 
 		assert.Equal(t, test.out, m)
 	}
+}
+
+func Test_transformCSV_BENCHMARK(t *testing.T) {
+	if os.Getenv("BENCHMARK") != "true" {
+		return
+	}
+	
+	// curl -LO https://s3.amazonaws.com/nyc-tlc/trip+data/yellow_tripdata_2021-04.csv
+	outTmp, err := ioutil.TempFile("", "")
+	defer os.Remove(outTmp.Name())
+	assert.Nil(t, err)
+
+
+	start := time.Now()
+	err = transformCSVFile("yellow_tripdata_2021-04.csv", outTmp, ',')
+	assert.Nil(t, err)
+
+	fmt.Printf("transform csv took %s\n", time.Since(start))
 }