--- a/runner/shape.go
+++ b/runner/shape.go
@@ -114,6 +114,7 @@ outer:
 		}
 
 		unique = append(unique, i)
+
 	}
 
 	return unique