--- a/runner/sql_test.go
+++ b/runner/sql_test.go
@@ -1,8 +1,12 @@
 package runner
 
 import (
-	"github.com/stretchr/testify/assert"
+	"io/ioutil"
+	"os"
 	"testing"
+
+	"github.com/google/uuid"
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_transformDM_getPanelCalls(t *testing.T) {
@@ -79,25 +83,25 @@ func Test_transformDM_getPanelCalls(t *testing.T) {
 	})
 }
 
-func Test_chunk(t *testing.T) {
-	a := []map[string]interface{}{
-		{"a": 1},
-		{"b": 2},
-		{"c": 3},
-		{"d": 4},
-		{"e": 5},
-		{"f": 6},
-		{"g": 7},
-	}
-	chunks := chunk(a, 3)
-	assert.Equal(t, len(chunks), 3)
-	assert.Equal(t, len(chunks[0]), 3)
-	assert.Equal(t, len(chunks[1]), 3)
-	assert.Equal(t, len(chunks[2]), 1)
-	assert.Equal(t, chunks[2][0], map[string]interface{}{
-		"g": 7,
-	})
-}
+// func Test_chunk(t *testing.T) {
+// 	a := []map[string]interface{}{
+// 		{"a": 1},
+// 		{"b": 2},
+// 		{"c": 3},
+// 		{"d": 4},
+// 		{"e": 5},
+// 		{"f": 6},
+// 		{"g": 7},
+// 	}
+// 	chunks := chunk(a, 3)
+// 	assert.Equal(t, len(chunks), 3)
+// 	assert.Equal(t, len(chunks[0]), 3)
+// 	assert.Equal(t, len(chunks[1]), 3)
+// 	assert.Equal(t, len(chunks[2]), 1)
+// 	assert.Equal(t, chunks[2][0], map[string]interface{}{
+// 		"g": 7,
+// 	})
+// }
 
 func Test_postgresMangleInsert(t *testing.T) {
 	assert.Equal(t,
@@ -160,4 +164,61 @@ func Test_getObjectAtPath(t *testing.T) {
 		res := getObjectAtPath(test.input, test.path)
 		assert.Equal(t, test.expected, res)
 	}
+}
+
+func Test_sqlIngest_BENCHMARK(t *testing.T) {
+	if os.Getenv("BENCHMARK") != "true" {
+		return
+	}
+
+	projectTmp, err := ioutil.TempFile("", "dsq-project")
+	assert.Nil(t, err)
+	defer os.Remove(projectTmp.Name())
+
+	project := &ProjectState{
+		Id: projectTmp.Name(),
+		Pages: []ProjectPage{
+			{
+				Panels: nil,
+			},
+		},
+	}
+
+	connector, tmp, err := MakeTmpSQLiteConnector()
+	assert.Nil(t, err)
+	defer os.Remove(tmp.Name())
+	project.Connectors = append(project.Connectors, *connector)
+
+	// curl -LO https://s3.amazonaws.com/nyc-tlc/trip+data/yellow_tripdata_2021-04.csv
+	// dsq yellow_tripdata_2021-04.csv > yellow_tripdata_2021-04.json
+	readFile := "yellow_tripdata_2021-04.json"
+	//readFile := "../testdata/allformats/userdata.json"
+
+	panelId := uuid.New().String()
+	s, err := ShapeFromFile(readFile, panelId, 10_000, 100)
+	assert.Nil(t, err)
+	project.Pages[0].Panels = append(project.Pages[0].Panels, PanelInfo{
+		ResultMeta: PanelResult{
+			Shape: *s,
+		},
+		Id:   panelId,
+		Name: uuid.New().String(),
+	})
+
+	panel2 := &PanelInfo{
+		Type:    DatabasePanel,
+		Content: "SELECT COUNT(1) FROM DM_getPanel(0)",
+		Id:      uuid.New().String(),
+		Name:    uuid.New().String(),
+		DatabasePanelInfo: &DatabasePanelInfo{
+			Database: DatabasePanelInfoDatabase{
+				ConnectorId: connector.Id,
+			},
+		},
+	}
+
+	err = EvalDatabasePanel(project, 0, panel2, func(projectId, panelId string) (chan map[string]interface{}, error) {
+		return loadJSONArrayFile(readFile)
+	})
+	assert.Nil(t, err)
 }